pyproject.toml
[project]
name = "ai-ppt-generator"
version = "0.1.0"
description = "AI PPT Generator using A2A and MCP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[sql]>=0.3.0",
    "click>=8.1.8",
    "fastmcp>=1.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.4.1",
    "mcp[cli]>=1.5.0",
    "python-pptx>=1.0.2",
    "pydantic>=2.11.4",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "numpy>=2.2.5",
    "nest-asyncio>=1.6.0",
]

[project.scripts]
ai-ppt = "ai_ppt:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_ppt"]

[project.optional-dependencies]
server = ["uvicorn>=0.34.0", "starlette>=0.45.3", "python-multipart>=0.0.20"]
